refactor: 구독 시 읽음 처리 서비스의 네이밍과 메서드명을 명확화
